% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastnrown.R
\name{fbind1}
\alias{fbind1}
\title{Bind two factors}
\usage{
fbind1(a, n)
}
\arguments{
\item{a}{factor}

\item{n}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
fbind1(a=c(1:10), n=3)
}
